refactor RacingCar.java, RacingCarSimulator.java: 패키지명 변경
패키지명을 컨벤션에 맞게 logic에서 domain으로 변경
